Finished refactoring main.py and gave most of the work to crawler_service.py and crawler_service_router.py
